Changed renderer.* to Graphics.* changed renderer functions from simple functional programming to OOP so now in classes. shaders.* now contains all shader related functions. Engine is also now using OOP rather than just functional. and updated main to reflect this. Same with makefile. Updated plan with some ramblings.
