+++ src/engine/engine.h
@@ -1,4 +1,15 @@
 #pragma once
-#include "graphics/renderobject.h"
-#include <vector>
-void engine();
+#include "models/components/renderobject.h"
+#include "models/Entity_data.h"
+
+class Engine
+{
+
+public:
+    Engine ( Entity_data entity_data ) ;
+    ~Engine ( ) ;
+    void run ();
+private:    
+    Entity_data entity_data;
+
+};
