+++ src/graphics/Graphics.h
@@ -1,0 +1,25 @@
+#pragma once
+#include <glad/glad.h>
+#include <GLFW/glfw3.h>
+
+class Graphics
+{
+
+private:
+    unsigned int shaderProgram;
+    GLFWwindow *window;
+
+    GLFWwindow* create_glfw_window ();
+
+public:
+    Graphics ( ) ;
+    ~Graphics ( ) ;
+    void Update ( ) ;
+
+};
+
+
+
+
+
+
