--- src/graphics/renderer.h
@@ -1,5 +1,0 @@
-#pragma once
-#include "renderobject.h"
-#include <vector>
-int renderer(std::vector<RenderObject> &objects);
-// init_buffers(float vertices[], std::size_t size);
