--- src/graphics/renderobject.cpp
@@ -1,33 +1,0 @@
-#include "renderobject.h"
-#include <glad/glad.h>
-#include <vector>
-
-RenderObject::RenderObject(std::vector<float> vertices) {
-  this->vertices = vertices;
-  this->VAO = 0;
-  this->VBO = 0;
-};
-
-int RenderObject::init_buffers() {
-  unsigned int VAO, VBO;
-  glGenVertexArrays(1, &VAO);
-  glBindVertexArray(VAO);
-
-  glGenBuffers(1, &VBO);
-
-  glBindBuffer(GL_ARRAY_BUFFER, VBO);
-  glBufferData(GL_ARRAY_BUFFER, vertices.size() * sizeof(float),
-               vertices.data(), GL_STATIC_DRAW);
-
-  glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 6 * sizeof(float), (void *)0);
-  glEnableVertexAttribArray(0);
-
-  glVertexAttribPointer(1, 3, GL_FLOAT, GL_FALSE, 6 * sizeof(float),
-                        (void *)(3 * sizeof(float)));
-  glEnableVertexAttribArray(1);
-
-  glBindVertexArray(0);
-  this->VAO = VAO;
-  this->VBO = VBO;
-  return 0;
-};
