+++ src/main.cpp
@@ -1,39 +1,37 @@
-#include "models/base/entity.h"
 #include <glad/glad.h>
+#include <unistd.h>
 #include <vector>
-
-#include "graphics/renderer.h"
-#include "models/plant/willow_tree.h"
 #include <GLFW/glfw3.h>
+#include "models/Entity_data.h"
 
-std::vector<float> vertices = {
+
+std::vector<float> vertices = 
+{
     -0.8f, -0.8f, 0.0f, 1.0f, 0.0f, 0.0f, // Vertex 1 (position and color)
     0.0f,  0.8f,  0.0f, 0.0f, 1.0f, 0.0f, // Vertex 2 (position and color)
     0.8f,  -0.8f, 0.0f, 0.0f, 0.0f, 1.0f  // Vertex 3 (position and color)
 };
-std::vector<float> vertices2 = {
+
+std::vector<float> vertices2 = 
+{
     -0.9f, 0.4f, 0.0f, 1.0f, 0.0f, 1.0f, // Vertex 1 (position and color)
     0.0f,  1.0f, 0.0f, 0.0f, 1.0f, 1.0f, // Vertex 2 (position and color)
     0.9f,  0.4f, 0.0f, 0.0f, 1.0f, 1.0f  // Vertex 3 (position and color)
 };
-std::vector<float> vertices3 = {
+
+std::vector<float> vertices3 = 
+{
     -0.6f, 0.0f,  0.0f, 1.0f, 1.0f, 0.0f, // Vertex 1 (position and color)
     0.0f,  -0.6f, 0.0f, 0.0f, 1.0f, 1.0f, // Vertex 2 (position and color)
     0.6f,  0.0f,  0.0f, 1.0f, 1.0f, 0.0f  // Vertex 3 (position and color)
 };
 
-std::vector<std::vector<float>> vec_objects = {vertices, vertices2, vertices3};
-int main() {
-  // Create renderObjects vector for each object we want to render
-  // pass to renderer
-  // willow_tree p = willow_tree{10, 10, 100, 1000, 50, vertices};
-  std::vector<Entity> x;
-  x.push_back();
+std::vector<std::vector<float>> vec_objects = { vertices, vertices2, vertices3 };
 
-  std::vector<RenderObject> objects;
-  for (auto &x : vec_objects) {
-    auto object = RenderObject{x};
-    objects.push_back(object);
-  }
-  renderer(objects);
+int main ( )
+{
+    Entity_data entity_data;
+    //set data.
+
 }
+
